LA.QSSI.14.0.r1-10300-qssi.0

[chrisl7]:
[1] - SKIP: BT: Dual Mode incremental patchset

* tag 'LA.QSSI.14.0.r1-10300-qssi.0' of https://git.codelinaro.org/clo/la/platform/vendor/qcom-opensource/system/bt:
  Handle LE link disconnected during BTM_SEC_STATE_ENCRYPTING
  Missing Null check for Client Interface
  Avoid duplicate hearing aid request
  BT: Dual Mode Incremental Changes (2).
  Revert "Select correct address to get remote name"
  BT: Dual Mode incremental patchset
  Fix potential null dereference after call to read_command_complete_header
  Fix an integer underflow in build_read_multi_rsp
  Revert "Revert "Fix wrong BR/EDR link key downgrades (P_256->P_192)""
  Revert "BT: SMP_RSP_TOUT happening due to LinkKey Refresh complete event"
  aidl: typecasting to avoid multiplication overflow
  BT: Reject RNR with invalid BD Address

+++ audio_hal_interface/aidl/le_audio_software.cc
@@ -150,7 +150,7 @@ bool LeAudioTransport::GetPresentationPosition(uint64_t* remote_delay_report_ns,
             << data_position_.tv_nsec
             << "s, delay report=" << remote_delay_report_ms_ << " msec.";
   if (remote_delay_report_ns != nullptr) {
-    *remote_delay_report_ns = (uint64_t)(remote_delay_report_ms_ * 1000000u);
+    *remote_delay_report_ns = (uint64_t)((uint64_t)remote_delay_report_ms_ * 1000000u);
   }
   if (total_bytes_processed != nullptr)
     *total_bytes_processed = (uint64_t)total_bytes_processed_;
